Implement code coverage for python and change provider to codecov (#326)
* fix(documentation) remove space between badges

* feature(pytests) add steadystate testcases

* feature(coverage) create merged report for cpp and py and change to codecov

* fix(codecov) remove codecov yaml

* fix(codecov) fix lcov command

* feature(codecov) simplify report upload

* fix(codecov) fix python coverage

* doc(codecov) add documentation to testCoverage.py + bugfixes

* fix(ci) add python coverage package in travis
