+++ .gitignore
@@ -187,3 +187,6 @@ python/.idea/workspace.xml
 python/testSpeedy.py
 
 python/test.txt
+
+tests/test/*
+coverage_py.xml
